feature/home/src/main/kotlin/com/ngapps/phototime/feature/home/HomeViewModel.kt
/*
 * Copyright 2024 NGApps Dev (https://github.com/ngapp-dev). All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.ngapps.phototime.feature.home

import androidx.lifecycle.ViewModel
import com.ngapps.phototime.core.data.util.SyncManager
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.flow.flowOf
import javax.inject.Inject

@HiltViewModel
class HomeViewModel @Inject constructor(
    syncManager: SyncManager,
) : ViewModel() {

    val uiState: Flow<HomeUiState> = flowOf()
}

sealed interface HomeUiState {
    data object Loading : HomeUiState

    data class Success(
        val homes: List<String>,
    ) : HomeUiState

    data object Empty : HomeUiState
}

